# backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Bike Auction Platform API
  description: |
    RESTful API for the Bike Auction Platform.

    ## Authentication
    - **Register/Login** → Receive a short-lived JWT Access Token (15 min) in the response body
      and a long-lived Refresh Token (7 days) as an `httpOnly`, `SameSite=Strict`, `Secure` cookie.
    - **Authenticated requests** → Pass the Access Token in the `Authorization: Bearer <token>` header.
    - **Token refresh** → `POST /api/auth/refresh` reads the cookie and returns a new Access Token.
    - **Real-time bidding** → Uses WebSocket (Socket.io) with token passed in the connection handshake.
      See the architecture document for WebSocket event contracts.

    ## Monetary Values
    All monetary fields are represented in **cents** (integer) to avoid floating-point precision issues.
    Example: `$1,250.00` → `125000`.
  version: "1.0.0"
  contact:
    name: Bike Auction Engineering
    email: eng@bikeauction.dev
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.bikeauction.dev
    description: Production

# ============================================================================
# Security Schemes
# ============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token (15 minutes). Obtain via POST /api/auth/login.

  # ==========================================================================
  # Reusable Schemas
  # ==========================================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: "Email is required"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # --- Auth ---
    RegisterRequest:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
          example: rider@example.com
        username:
          type: string
          minLength: 3
          maxLength: 100
          example: speed_rider
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          example: "S3cur3P@ss!"
        role:
          type: string
          enum: [buyer, seller]
          default: buyer
          description: "Admin role cannot be self-assigned."
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: rider@example.com
        password:
          type: string
          format: password
          example: "S3cur3P@ss!"
    AuthResponse:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
          description: Short-lived JWT (15 min). Refresh token is set as httpOnly cookie.
          example: eyJhbGciOiJIUzI1NiIs...
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        role:
          type: string
          enum: [buyer, seller, admin]
        isSellerVerified:
          type: boolean
        hasCardOnFile:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # --- Motorcycle ---
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created. Refresh token set as httpOnly cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=abc123; HttpOnly; Secure; SameSite=Strict; Path=/api/auth; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error (missing fields, weak password)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email or username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful. Refresh token set as httpOnly cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=abc123; HttpOnly; Secure; SameSite=Strict; Path=/api/auth; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using httpOnly cookie
      operationId: refreshToken
      description: |
        Reads the `refreshToken` from the httpOnly cookie.
        Returns a new access token and rotates the refresh token.
      responses:
        "200":
          description: New access token issued. New refresh token cookie set.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        "401":
          description: Refresh token missing, expired, or revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout and revoke refresh token
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out. Refresh token cookie cleared.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --------------------------------------------------------------------------
  # Motorcycles
  # --------------------------------------------------------------------------
...